refactor: Update spacing and button styles
Update the spacing between sections and items in the sizes.dart file. Also, add new button sizes and elevation.

In the app.dart file, replace the WelcomeScreen with the LoginPage.

In the colors.dart file, add new shades of grey.

In the theme.dart file, import and use the new elevated button theme.

In the strings.dart file, update the text for the forgot password and create account options.

In the text_theme.dart file, update the text colors to use the primary color.